+++ {{ module_name }}/src/{{ module_name }}/{{ module_name }}/backend_webserver/settings/__init__.py
@@ -1,0 +1,5 @@
+"""Settings endpoints for {{ module_name }}."""
+
+from .router import router as settings_router
+
+__all__ = ["settings_router"]